[Ang] Completed frontend AB All Module
This module compares all entries in the database with the current input string by the user.
It emits an event of request, then on any response it adds into the array.

Downsides being, all results will be added into RAM, making comparisons taxing on the users' computer on the long run.

Plans to cache and serve  on demand are being planned out, but will not be done until works for all workshop modules are complete.